requirement.txt
flask
flask-socketio
yt-dlp
faster-whisper
openai
python-socketio
sqlalchemy
python-dotenv
pyyaml
torch
torchvision
torchaudio
